MetisAgent2/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Sock==0.7.0
Flask-SocketIO==5.3.6
requests==2.31.0
python-dotenv==1.0.0
chromadb==0.4.15
sentence-transformers==2.2.2
openai==1.3.0
Pillow==10.0.0
cryptography==41.0.4
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.1
instagrapi==2.1.2